added main empty and add tests
